エディタ側の追加処理のバグを修正 再コンパイルボタンを追加
