+++ project/engine/BuildInfo.h
@@ -1,5 +1,5 @@
 #pragma once 
-#define BUILD_COMMIT "456f05e4" 
-#define BUILD_BRANCH "develop" 
-#define BUILD_DATE "2026/04/27" 
-#define BUILD_TIME "17:29:27.49" 
+#define BUILD_COMMIT "abed13f7" 
+#define BUILD_BRANCH "feat-C#" 
+#define BUILD_DATE "2026/04/28" 
+#define BUILD_TIME "16:35:46.91"
